Merge ProcedureManager mDefinitions and mReferences maps (#622)
This is a WIP on functions.

I replaced the two maps with one, in an effort to reduce the number of parallel keys I need to maintain.

Future step is to do the same with the NameManager, making it a map-like structure to ProcedureBlocks or VariableInfo objects. This also centralizes the canonization of names, and reduces the "surface area" of the name synchronization problem seen during procedure rename.

Also, much more ProcedureManager JavaDocs I missed on the last pass.